Grouping Menu events in MenuEventArgs.cs file.
